requirements.txt
Flask
FastApi
Flask-SQLAlchemy
Jinja2
requests
python-dotenv
sqlalchemy